change /users/register and /users/login base path to /account/
