+++ LeaderboardBackend.Test/TestApi/Extensions/UserApiExtensions.cs
@@ -15,7 +15,7 @@ internal static class UserApiExtensions
     )
     {
         return await client.Post<UserViewModel>(
-            "/api/users/register",
+            "/api/account/register",
             new()
             {
                 Body = new RegisterRequest()
@@ -35,7 +35,7 @@ internal static class UserApiExtensions
     )
     {
         return await apiClient.Post<LoginResponse>(
-            "/api/users/login",
+            "/api/account/login",
             new()
             {
                 Body = new LoginRequest() { Email = email, Password = password, }
@@ -46,7 +46,7 @@ internal static class UserApiExtensions
     public static async Task<LoginResponse> LoginAdminUser(this TestApiClient apiClient)
     {
         return await apiClient.Post<LoginResponse>(
-            "/api/users/login",
+            "/api/account/login",
             new()
             {
                 Body = new LoginRequest()
